Update actual text read by screen readers (#100)
* Update actual text read by screen readers

Maybe that was what screen readers said in the early stage of the excol component. Now both JAWS and NVDA say the phrase that I use in this pull request. NVDA, in addition, adds the heading level information,and JAWS currently omits it. I don't have a Mac currently but I had this component tested by a Voice Over user and it outputs more or less the same information.

Co-authored-by: Eric Eggert <mail@yatil.net>